LRCI-6021 Move methods to get a downstream build json object to BaseBuildReport
